refactor: move `CancelTransactionModal`
