+++ packages/wallets/src/features/cashier/components/CancelTransactionModal/CancelTransactionModal.scss
